[INLONG-5562][Manager] Move Workflow UT to a new class
